requirements.txt
Babel==2.12.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.7
fonttools==4.42.1
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
pywin32==306
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.2.0
tk==0.1.0
tqdm==4.66.1
tzdata==2023.3
zipp==3.16.2
